move layout actions to their own component
